fix bullet prespawn
