+++ examples/avian_physics/Cargo.toml
@@ -7,28 +7,36 @@ publish = false
 
 [features]
 default = [
-  "lightyear_examples_common/server",
-  "lightyear_examples_common/client",
+    "lightyear_examples_common/websocket"
+]
+client = [
+    "lightyear_examples_common/client",
+    "gui",
+]
+server = [
+    "lightyear_examples_common/server"
+]
+gui = [
+    "lightyear_examples_common/gui"
 ]
 metrics = ["lightyear/metrics", "dep:metrics-exporter-prometheus"]
 
 [dependencies]
 lightyear_examples_common.workspace = true
 lightyear = { workspace = true, features = [
-  "steam",
-  "websocket",
-  "leafwing",
-  "avian2d",
+    "steam",
+    "websocket",
+    "leafwing",
+    "avian2d",
 ] }
 
-bevy_screen_diagnostics.workspace = true
 leafwing-input-manager.workspace = true
 avian2d = { workspace = true, features = [
-  "2d",
-  "f32",
-  "parry-f32",
-  "parallel",
-  "serialize",
+    "2d",
+    "f32",
+    "parry-f32",
+    "parallel",
+    "serialize",
 ] }
 serde.workspace = true
 anyhow.workspace = true
