+++ examples/avian_physics/src/main.rs
@@ -1,8 +1,6 @@
 #![allow(unused_imports)]
 #![allow(unused_variables)]
 #![allow(dead_code)]
-use crate::client::ExampleClientPlugin;
-use crate::server::ExampleServerPlugin;
 use crate::shared::SharedPlugin;
 use bevy::prelude::*;
 use lightyear::prelude::client::PredictionConfig;
@@ -10,11 +8,15 @@ use lightyear_examples_common::app::{Apps, Cli};
 use lightyear_examples_common::settings::{read_settings, Settings};
 use serde::{Deserialize, Serialize};
 
+#[cfg(feature = "client")]
 mod client;
 mod protocol;
+
+#[cfg(feature = "gui")]
+mod renderer;
+#[cfg(feature = "server")]
 mod server;
 mod shared;
-
 fn main() {
     let cli = Cli::default();
     let settings_str = include_str!("../assets/settings.ron");
@@ -26,17 +28,21 @@ fn main() {
     apps.update_lightyear_client_config(|config| {
         config.prediction.minimum_input_delay_ticks = settings.input_delay_ticks;
         config.prediction.correction_ticks_factor = settings.correction_ticks_factor;
-    })
-    // add `ClientPlugins` and `ServerPlugins` plugin groups
-    .add_lightyear_plugins()
-    // add our plugins
-    .add_user_client_plugin(ExampleClientPlugin)
-    .add_user_server_plugin(ExampleServerPlugin {
+    });
+
+    apps.add_lightyear_plugins();
+    apps.add_user_shared_plugin(SharedPlugin);
+    #[cfg(feature = "client")]
+    apps.add_user_client_plugin(client::ExampleClientPlugin);
+    #[cfg(feature = "server")]
+    apps.add_user_server_plugin(server::ExampleServerPlugin {
         predict_all: settings.predict_all,
-    })
-    .add_user_shared_plugin(SharedPlugin {
+    });
+    #[cfg(feature = "gui")]
+    apps.add_user_renderer_plugin(crate::renderer::ExampleRendererPlugin {
         show_confirmed: settings.show_confirmed,
     });
+
     // run the app
     apps.run();
 }
