+++ examples/avian_physics/src/shared.rs
@@ -2,9 +2,7 @@ use avian2d::prelude::*;
 use bevy::color::palettes::css;
 use bevy::diagnostic::LogDiagnosticsPlugin;
 use bevy::prelude::*;
-use bevy::render::RenderPlugin;
 use bevy::utils::Duration;
-use bevy_screen_diagnostics::{Aggregate, ScreenDiagnostics, ScreenDiagnosticsPlugin};
 use leafwing_input_manager::prelude::ActionState;
 use tracing::Level;
 
@@ -17,45 +15,22 @@ use lightyear::transport::io::IoDiagnosticsPlugin;
 use lightyear_examples_common::shared::FIXED_TIMESTEP_HZ;
 
 use crate::protocol::*;
-const MAX_VELOCITY: f32 = 200.0;
+pub(crate) const MAX_VELOCITY: f32 = 200.0;
 const WALL_SIZE: f32 = 350.0;
 
 #[derive(Clone)]
-pub struct SharedPlugin {
-    pub(crate) show_confirmed: bool,
-}
+pub struct SharedPlugin;
 
 impl Plugin for SharedPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(ProtocolPlugin);
-        if app.is_plugin_added::<RenderPlugin>() {
-            app.add_systems(Startup, init_camera);
-
-            // draw after interpolation is done
-            app.add_systems(
-                PostUpdate,
-                draw_elements
-                    .after(InterpolationSet::Interpolate)
-                    .after(PredictionSet::VisualCorrection),
-            );
-            if self.show_confirmed {
-                app.add_systems(
-                    PostUpdate,
-                    draw_confirmed_shadows
-                        .after(InterpolationSet::Interpolate)
-                        .after(PredictionSet::VisualCorrection),
-                );
-            }
-            app.add_plugins(LogDiagnosticsPlugin {
-                filter: Some(vec![
-                    IoDiagnosticsPlugin::BYTES_IN,
-                    IoDiagnosticsPlugin::BYTES_OUT,
-                ]),
-                ..default()
-            });
-            // app.add_systems(Startup, setup_diagnostic);
-            // app.add_plugins(ScreenDiagnosticsPlugin::default());
-        }
+        app.add_plugins(LogDiagnosticsPlugin {
+            filter: Some(vec![
+                IoDiagnosticsPlugin::BYTES_IN,
+                IoDiagnosticsPlugin::BYTES_OUT,
+            ]),
+            ..default()
+        });
         // bundles
         app.add_systems(Startup, init);
 
@@ -79,50 +54,12 @@ impl Plugin for SharedPlugin {
     }
 }
 
-// fn setup_diagnostic(mut onscreen: ResMut<ScreenDiagnostics>) {
-//     onscreen
-//         .add(
-//             "Rollbacks".to_string(),
-//             PredictionDiagnosticsPlugin::ROLLBACKS,
-//         )
-//         .aggregate(Aggregate::Value)
-//         .format(|v| format!("{v:.0}"));
-//     onscreen
-//         .add(
-//             "Rollback ticks".to_string(),
-//             PredictionDiagnosticsPlugin::ROLLBACK_TICKS,
-//         )
-//         .aggregate(Aggregate::Value)
-//         .format(|v| format!("{v:.0}"));
-//     onscreen
-//         .add(
-//             "RB depth".to_string(),
-//             PredictionDiagnosticsPlugin::ROLLBACK_DEPTH,
-//         )
-//         .aggregate(Aggregate::Value)
-//         .format(|v| format!("{v:.1}"));
-//     // screen diagnostics twitches due to layout change when a metric adds or removes
-//     // a digit, so pad these metrics to 3 digits.
-//     onscreen
-//         .add("KB_in".to_string(), IoDiagnosticsPlugin::BYTES_IN)
-//         .aggregate(Aggregate::Average)
-//         .format(|v| format!("{v:0>3.0}"));
-//     onscreen
-//         .add("KB_out".to_string(), IoDiagnosticsPlugin::BYTES_OUT)
-//         .aggregate(Aggregate::Average)
-//         .format(|v| format!("{v:0>3.0}"));
-// }
-
 // Generate pseudo-random color from id
 pub(crate) fn color_from_id(client_id: ClientId) -> Color {
     let h = (((client_id.to_bits().wrapping_mul(30)) % 360) as f32) / 360.0;
     let s = 1.0;
     let l = 0.5;
     Color::hsl(h, s, l)
-}
-
-fn init_camera(mut commands: Commands) {
-    commands.spawn(Camera2dBundle::default());
 }
 
 pub(crate) fn init(mut commands: Commands) {
@@ -229,62 +166,6 @@ pub(crate) fn log() {
     trace!("run physics schedule!");
 }
 
-/// System that draws the outlines of confirmed entities, with lines to the centre of their predicted location.
-pub(crate) fn draw_confirmed_shadows(
-    mut gizmos: Gizmos,
-    confirmed_q: Query<(&Position, &Rotation, &LinearVelocity, &Confirmed), With<PlayerId>>,
-    predicted_q: Query<&Position, With<PlayerId>>,
-) {
-    for (position, rotation, velocity, confirmed) in confirmed_q.iter() {
-        let speed = velocity.length() / MAX_VELOCITY;
-        let ghost_col = css::GRAY.with_alpha(speed);
-        gizmos.rect_2d(
-            Isometry2d {
-                rotation: Rot2 {
-                    sin: rotation.sin,
-                    cos: rotation.cos,
-                },
-                translation: Vec2::new(position.x, position.y),
-            },
-            Vec2::ONE * PLAYER_SIZE,
-            ghost_col,
-        );
-        if let Some(e) = confirmed.predicted {
-            if let Ok(pos) = predicted_q.get(e) {
-                gizmos.line_2d(**position, **pos, ghost_col);
-            }
-        }
-    }
-}
-
-/// System that draws the player's boxes and cursors
-pub(crate) fn draw_elements(
-    mut gizmos: Gizmos,
-    players: Query<(&Position, &Rotation, &ColorComponent), (Without<Confirmed>, With<PlayerId>)>,
-    balls: Query<(&Position, &ColorComponent), (Without<Confirmed>, With<BallMarker>)>,
-    walls: Query<(&Wall, &ColorComponent), (Without<BallMarker>, Without<PlayerId>)>,
-) {
-    for (position, rotation, color) in &players {
-        gizmos.rect_2d(
-            Isometry2d {
-                rotation: Rot2 {
-                    sin: rotation.sin,
-                    cos: rotation.cos,
-                },
-                translation: Vec2::new(position.x, position.y),
-            },
-            Vec2::ONE * PLAYER_SIZE,
-            color.0,
-        );
-    }
-    for (position, color) in &balls {
-        gizmos.circle_2d(Vec2::new(position.x, position.y), BALL_SIZE, color.0);
-    }
-    for (wall, color) in &walls {
-        gizmos.line_2d(wall.start, wall.end, color.0);
-    }
-}
-
 // Wall
 #[derive(Bundle)]
 pub(crate) struct WallBundle {
@@ -296,8 +177,8 @@ pub(crate) struct WallBundle {
 
 #[derive(Component)]
 pub(crate) struct Wall {
-    start: Vec2,
-    end: Vec2,
+    pub(crate) start: Vec2,
+    pub(crate) end: Vec2,
 }
 
 impl WallBundle {
