+++ examples/priority/src/main.rs
@@ -1,16 +1,18 @@
 #![allow(unused_imports)]
 #![allow(unused_variables)]
 #![allow(dead_code)]
-use crate::client::ExampleClientPlugin;
-use crate::server::ExampleServerPlugin;
-use crate::shared::SharedPlugin;
 use bevy::prelude::*;
 use lightyear::prelude::server::PacketConfig;
 use lightyear_examples_common::app::{Apps, Cli};
 use lightyear_examples_common::settings::{read_settings, Settings};
 
+#[cfg(feature = "client")]
 mod client;
 mod protocol;
+
+#[cfg(feature = "gui")]
+mod renderer;
+#[cfg(feature = "server")]
 mod server;
 mod shared;
 
@@ -27,13 +29,16 @@ fn main() {
             .enable_bandwidth_cap()
             // we can set the max bandwidth to 56 KB/s
             .with_send_bandwidth_bytes_per_second_cap(3000);
-    })
-    // add the `ClientPlugins` and `ServerPlugins` plugin groups
-    .add_lightyear_plugins();
-    // add our plugins
-    apps.add_user_client_plugin(ExampleClientPlugin);
-    apps.add_user_server_plugin(ExampleServerPlugin);
-    apps.add_user_shared_plugin(SharedPlugin);
+    });
+
+    apps.add_lightyear_plugins();
+    apps.add_user_shared_plugin(shared::SharedPlugin);
+    #[cfg(feature = "client")]
+    apps.add_user_client_plugin(client::ExampleClientPlugin);
+    #[cfg(feature = "server")]
+    apps.add_user_server_plugin(server::ExampleServerPlugin);
+    #[cfg(feature = "gui")]
+    apps.add_user_renderer_plugin(renderer::ExampleRendererPlugin);
     // run the app
     apps.run();
 }
