+++ examples/priority/src/shared.rs
@@ -21,13 +21,6 @@ pub struct SharedPlugin;
 impl Plugin for SharedPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(ProtocolPlugin);
-        if app.is_plugin_added::<RenderPlugin>() {
-            app.add_systems(Startup, init);
-            app.add_systems(Update, (draw_players, draw_props));
-            // diagnostics
-            // app.add_systems(Startup, setup_diagnostic);
-            // app.add_plugins(ScreenDiagnosticsPlugin::default());
-        }
 
         // movement
         app.add_systems(FixedUpdate, player_movement);
@@ -37,17 +30,6 @@ impl Plugin for SharedPlugin {
 fn init(mut commands: Commands) {
     commands.spawn(Camera2d::default());
 }
-
-// fn setup_diagnostic(mut onscreen: ResMut<ScreenDiagnostics>) {
-//     onscreen
-//         .add("KB/S in".to_string(), IoDiagnosticsPlugin::BYTES_IN)
-//         .aggregate(Aggregate::Average)
-//         .format(|v| format!("{v:.2}"));
-//     onscreen
-//         .add("KB/s out".to_string(), IoDiagnosticsPlugin::BYTES_OUT)
-//         .aggregate(Aggregate::Average)
-//         .format(|v| format!("{v:.2}"));
-// }
 
 /// Read client inputs and move players
 pub(crate) fn player_movement(
@@ -65,51 +47,6 @@ pub(crate) fn player_movement(
         }
         if input.pressed(&Inputs::Right) {
             position.x += MOVE_SPEED;
-        }
-    }
-}
-
-/// System that draws the player
-/// The components should be replicated from the server to the client
-/// This time we will only draw the predicted/interpolated entities
-pub(crate) fn draw_players(
-    mut gizmos: Gizmos,
-    players: Query<(&Position, &PlayerColor), Without<Confirmed>>,
-) {
-    for (position, color) in &players {
-        gizmos.rect(
-            Isometry3d::from_translation(Vec3::new(position.x, position.y, 0.0)),
-            Vec2::ONE * 50.0,
-            color.0,
-        );
-    }
-}
-
-/// System that draws the props
-pub(crate) fn draw_props(mut gizmos: Gizmos, props: Query<(&Position, &Shape)>) {
-    for (position, shape) in props.iter() {
-        match shape {
-            Shape::Circle => {
-                gizmos.circle_2d(*position.deref(), PROP_SIZE, GREEN);
-            }
-            Shape::Triangle => {
-                gizmos.linestrip_2d(
-                    vec![
-                        *position.deref() + Vec2::new(0.0, PROP_SIZE),
-                        *position.deref() + Vec2::new(PROP_SIZE, -PROP_SIZE),
-                        *position.deref() + Vec2::new(-PROP_SIZE, -PROP_SIZE),
-                        *position.deref() + Vec2::new(0.0, PROP_SIZE),
-                    ],
-                    RED,
-                );
-            }
-            Shape::Square => {
-                gizmos.rect_2d(
-                    Isometry2d::from_translation(*position.deref()),
-                    Vec2::splat(PROP_SIZE * 2.0),
-                    BLUE,
-                );
-            }
         }
     }
 }
