+++ examples/replication_groups/src/main.rs
@@ -1,15 +1,17 @@
 #![allow(unused_imports)]
 #![allow(unused_variables)]
 #![allow(dead_code)]
-use crate::client::ExampleClientPlugin;
-use crate::server::ExampleServerPlugin;
-use crate::shared::SharedPlugin;
 use bevy::prelude::*;
 use lightyear_examples_common::app::{Apps, Cli};
 use lightyear_examples_common::settings::{read_settings, Settings};
 
+#[cfg(feature = "client")]
 mod client;
 mod protocol;
+
+#[cfg(feature = "gui")]
+mod renderer;
+#[cfg(feature = "server")]
 mod server;
 mod shared;
 
@@ -20,12 +22,16 @@ fn main() {
     // build the bevy app (this adds common plugin such as the DefaultPlugins)
     // and returns the `ClientConfig` and `ServerConfig` so that we can modify them if needed
     let mut apps = Apps::new(settings, cli, env!("CARGO_PKG_NAME").to_string());
-    // add the `ClientPlugins` and `ServerPlugins` plugin groups
+
     apps.add_lightyear_plugins();
-    // add our plugins
-    apps.add_user_client_plugin(ExampleClientPlugin);
-    apps.add_user_server_plugin(ExampleServerPlugin);
-    apps.add_user_shared_plugin(SharedPlugin);
+    apps.add_user_shared_plugin(shared::SharedPlugin);
+    #[cfg(feature = "client")]
+    apps.add_user_client_plugin(client::ExampleClientPlugin);
+    #[cfg(feature = "server")]
+    apps.add_user_server_plugin(server::ExampleServerPlugin);
+    #[cfg(feature = "gui")]
+    apps.add_user_renderer_plugin(renderer::ExampleRendererPlugin);
+
     // run the app
     apps.run();
 }
