+++ examples/replication_groups/src/shared.rs
@@ -23,10 +23,6 @@ impl Plugin for SharedPlugin {
         app.register_type::<PlayerColor>();
         app.register_type::<TailPoints>();
         app.register_type::<TailLength>();
-        if app.is_plugin_added::<RenderPlugin>() {
-            app.add_systems(Startup, init);
-            app.add_systems(Update, draw_snakes);
-        }
     }
 }
 
@@ -91,41 +87,5 @@ pub(crate) fn shared_tail_behaviour(
         // Update the back
         // remove the back points that are above the length
         points.shorten_back(parent_position.0, length.0);
-    }
-}
-
-/// System that draws the boxes of the player positions.
-/// The components should be replicated from the server to the client
-pub(crate) fn draw_snakes(
-    mut gizmos: Gizmos,
-    players: Query<(&PlayerPosition, &PlayerColor), Without<Confirmed>>,
-    tails: Query<(&PlayerParent, &TailPoints), Without<Confirmed>>,
-) {
-    for (parent, points) in tails.iter() {
-        debug!("drawing snake with parent: {:?}", parent.0);
-        let Ok((position, color)) = players.get(parent.0) else {
-            error!("Tail entity has no parent entity!");
-            continue;
-        };
-        // draw the head
-        gizmos.rect(
-            Isometry3d::from_translation(Vec3::new(position.x, position.y, 0.0)),
-            Vec2::ONE * 20.0,
-            color.0,
-        );
-        // draw the first line
-        gizmos.line_2d(position.0, points.0.front().unwrap().0, color.0);
-        if position.0.x != points.0.front().unwrap().0.x
-            && position.0.y != points.0.front().unwrap().0.y
-        {
-            debug!("DIAGONAL");
-        }
-        // draw the rest of the lines
-        for (start, end) in points.0.iter().zip(points.0.iter().skip(1)) {
-            gizmos.line_2d(start.0, end.0, color.0);
-            if start.0.x != end.0.x && start.0.y != end.0.y {
-                debug!("DIAGONAL");
-            }
-        }
     }
 }
